+++ packages/flagship/src/types.ts
@@ -169,6 +169,13 @@ export interface Config {
 
   targetedDevices?: TargetedDevices;
 
+  /**
+   * Array of environment names to hide from the Environment Switcher
+   *
+   * @example ['store']
+   */
+  hiddenEnvs?: string[];
+
   webPath?: string;
   webTitle?: string;
   webScriptInjectHeader?: string;
@@ -217,11 +224,23 @@ export interface IOSConfig {
    */
   extensions?: IOSExtension[];
   shortVersion?: string;
+  frameworks?: FrameworksConfig[];
 }
 
 export interface PodsConfig {
   sources?: string[];
   newPods?: string[];
+}
+
+export interface FrameworksConfig {
+  /**
+   * iOS Framework to be added to project
+   */
+  framework: string;
+  /**
+   * iOS Framework directory path from project root source
+   */
+  frameworkPath?: string;
 }
 
 export interface NPMPackageConfig {