+++ packages/fsapp/src/beta-app/app/context/index.ts
@@ -1,2 +1,3 @@
 export * from './api.context';
 export * from './app.context';
+export * from './store.context';