+++ packages/fsapp/src/beta-app/router/router.tsx
@@ -28,6 +28,8 @@ import {
 import { FSRouterBase } from './router.base';
 import { trackView } from './utils';
 
+export { NAVIGATOR_TOKEN } from './context/navigator.context';
+
 // This is a hack. I am not happy about having to do this hack.
 // But it is required for Android. If no components are registered
 // Synchronously then errors are thrown and touch responses are eaten.