+++ packages/fscomponents/src/index.ts
@@ -1,5 +1,6 @@
 export { default as applyDefaults, ProjectWrapperComponent } from './lib/applyDefaults';
 export { default as ApplePay } from './lib/wallets/ApplePay';
+export * from './hooks/grid-chunks.hook';
 export * from './components';
 export * from './lib/color';
 export * from './lib/style';