+++ packages/fsengagement/src/inboxblocks/IconTextBlock.tsx
@@ -1,4 +1,4 @@
-import React, { Component } from 'react';
+import React, { Component, useContext } from 'react';
 import {
   Image,
   ImageStyle,
@@ -13,6 +13,7 @@ import { TextBlock } from './TextBlock';
 import { ImageBlock } from './ImageBlock';
 import { CTABlock } from './CTABlock';
 import DividerBlock from './DividerBlock';
+import { EngagementContext } from '../lib/contexts';
 
 const images = {
   rightArrow: require('../../assets/images/rightArrow.png'),
@@ -65,7 +66,7 @@ export interface IconTextProps {
 }
 export type ArrowTypes = 'rightArrow' | 'rightBlockArrow' | 'rightCategoryArrow';
 
-export default class IconTextBlock extends Component<IconTextProps> {
+class IconTextBlock extends Component<IconTextProps & { context: any }> {
   static contextTypes: any = {
     handleAction: PropTypes.func
   };
@@ -73,7 +74,7 @@ export default class IconTextBlock extends Component<IconTextProps> {
     if (!link) {
       return;
     }
-    const { handleAction } = this.context;
+    const { handleAction } = this.props.context;
     handleAction({
       type: 'deep-link',
       value: link
@@ -158,3 +159,8 @@ export default class IconTextBlock extends Component<IconTextProps> {
     );
   }
 }
+
+export default (props: IconTextProps) => {
+  const context = useContext(EngagementContext);
+  return <IconTextBlock {...props} context={context} />;
+};